% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bubblegraph-method.R
\docType{methods}
\name{bubblegraph}
\alias{bubblegraph}
\alias{bubblegraph,data.frame-method}
\alias{bubblegraph,matrix-method}
\title{bubblegraph visualisation}
\usage{
bubblegraph(object, ...)

\S4method{bubblegraph}{matrix}(object, col = "darkgrey", rex = 1,
  leftMargin = 1.2, bottomMargin = 1.6, cex = 1, font = 1, ...)

\S4method{bubblegraph}{data.frame}(object, col = "darkgrey", rex = 1,
  leftMargin = 1.2, bottomMargin = 1.6, cex = 1, font = 1, ...)
}
\arguments{
\item{object}{a data.frame}

\item{...}{further parameters that will be passed to the plot function}

\item{col}{color, either data.frame or vector}

\item{rex}{a radius expansion factor to control the bubble size}

\item{leftMargin}{adjust left margin of the plot}

\item{bottomMargin}{adjust bottom margin of the plot}

\item{cex}{character expansion factor}

\item{font}{set font parameter from par}
}
\value{
a nice plot, hopefully
}
\description{
visualise the distribution of counts/frequencies
}
\examples{
\dontrun{
library(polmineR)
bt <- partition("PLPRBTTXT", def=list(text_lp="17", text_type="speech"))
dist <- dispersion(bt, "Energiewende", c("text_year", "text_party"))
bubblegraph(slot(dist, "abs"))
}
}

