% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linechart-method.R
\docType{methods}
\name{linechart}
\alias{linechart}
\alias{linechart,data.frame-method}
\title{Line chart method}
\usage{
linechart(object, ...)

\S4method{linechart}{data.frame}(object, xlab = "", ylab = "", main = "",
  col = NULL, ymax = NULL)
}
\arguments{
\item{object}{data to be presented (data.frame)}

\item{...}{further arguments}

\item{xlab}{xlab (cp. plot)}

\item{ylab}{ylab (cp. plot)}

\item{main}{main title (cp.plot)}

\item{col}{color for rows}

\item{ymax}{maximum y value for chart}
}
\value{
output is a plot
}
\description{
Draw a line chart
}
\details{
Suitable for time series data. If time series data is provided, dates
should be in columns
}
\examples{
\dontrun{
library(polmineR)
bt <- partition("PLPRBTTXT", def=list(text_lp="17", text_type="speech"))
dist <- dispersion(bt, "Energiewende", c("text_year", "text_party"))
linechart(slot(dist, "abs"))
}
}
\author{
Andreas Blaette
}

